+++ Runtime/ARKitApi.cs
@@ -59,6 +59,9 @@ namespace UnityEngine.XR.ARKit
         [DllImport("__Internal", EntryPoint="UnityARKit_Version_AtLeast13_0")]
         public static extern bool AtLeast13_0();
 
+        [DllImport("__Internal", EntryPoint="UnityARKit_Version_AtLeast16_0")]
+        public static extern bool AtLeast16_0();
+
         [DllImport("__Internal", EntryPoint="UnityARKit_EnsureRootViewIsSetup")]
         public static extern void EnsureRootViewIsSetup();
 
@@ -145,6 +148,7 @@ namespace UnityEngine.XR.ARKit
         public static bool AtLeast11_3() => false;
         public static bool AtLeast12_0() => false;
         public static bool AtLeast13_0() => false;
+        public static bool AtLeast16_0() => false;
 #endif
     }
 }